+++ src/OpenTelemetry.Exporter.Geneva/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Add support for exporting `UpDownCounter` and `ObservableUpDownCounter`.
   [#685](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/685)
+
 ## 1.4.0-beta.2
 
 Released 2022-Oct-17